refactor(engine): decompose context_impl/options.rs (SPL-29 #2457)
Splits the 1114-line context_impl/options.rs into five focused submodules
under context_impl/options/, grouped by option family:

- metadata.rs  - MetadataOptions snapshot, ACL/xattr/numeric-id accessors
- transfer.rs  - link/sparse/append/compress/checksum/partial accessors
- dirs.rs      - directory/path accessors plus prepare_parent_directory
- filter.rs    - filter program access and allows/allows_deletion predicates
- batch.rs     - batch writer flow plus the batch_entry_compare helper

Pure refactor: every public and pub(super)/pub(crate) item retains its
original signature, visibility, and call site. context.rs swaps a single
include! for five, preserving the inline-merge with CopyContext.
