+++ crates/engine/src/local_copy/context_impl/options/batch.rs
@@ -1,521 +1,4 @@
-use std::num::NonZeroU32;
-
 impl<'a> CopyContext<'a> {
-    /// Builds a [`MetadataOptions`] snapshot from the current copy options.
-    pub(super) fn metadata_options(&self) -> MetadataOptions {
-        MetadataOptions::new()
-            .preserve_owner(self.options.preserve_owner())
-            .preserve_group(self.options.preserve_group())
-            .preserve_executability(self.options.preserve_executability())
-            .preserve_permissions(self.options.preserve_permissions())
-            .preserve_times(self.options.preserve_times())
-            .preserve_atimes(self.options.preserve_atimes())
-            .preserve_crtimes(self.options.preserve_crtimes())
-            .numeric_ids(self.options.numeric_ids_enabled())
-            .fake_super(self.options.fake_super_enabled())
-            .with_owner_override(self.options.owner_override())
-            .with_group_override(self.options.group_override())
-            .with_chmod(self.options.chmod().cloned())
-            .with_user_mapping(self.options.user_mapping().cloned())
-            .with_group_mapping(self.options.group_mapping().cloned())
-    }
-
-    pub(super) const fn copy_links_enabled(&self) -> bool {
-        self.options.copy_links_enabled()
-    }
-
-    pub(super) const fn links_enabled(&self) -> bool {
-        self.options.links_enabled()
-    }
-
-    pub(super) const fn copy_unsafe_links_enabled(&self) -> bool {
-        self.options.copy_unsafe_links_enabled()
-    }
-
-    pub(super) const fn safe_links_enabled(&self) -> bool {
-        self.options.safe_links_enabled()
-    }
-
-    pub(super) const fn munge_links_enabled(&self) -> bool {
-        self.options.munge_links_enabled()
-    }
-
-    /// Sets the safety depth offset.  Call with `1` before entering a
-    /// whole-directory copy (no trailing slash), `0` for a contents copy.
-    pub(super) fn set_safety_depth_offset(&mut self, offset: usize) {
-        self.safety_depth_offset = offset;
-    }
-
-    /// Strips the transfer-root prefix from a relative path, producing a
-    /// path suitable for `symlink_target_is_safe`.
-    pub(super) fn strip_safety_prefix<'p>(&self, relative: &'p Path) -> &'p Path {
-        if self.safety_depth_offset == 0 {
-            return relative;
-        }
-        let mut components = relative.components();
-        for _ in 0..self.safety_depth_offset {
-            components.next();
-        }
-        components.as_path()
-    }
-
-    pub(super) const fn copy_dirlinks_enabled(&self) -> bool {
-        self.options.copy_dirlinks_enabled()
-    }
-
-    pub(super) const fn keep_dirlinks_enabled(&self) -> bool {
-        self.options.keep_dirlinks_enabled()
-    }
-
-    pub(super) const fn whole_file_enabled(&self) -> bool {
-        self.options.whole_file_enabled()
-    }
-
-    pub(super) const fn open_noatime_enabled(&self) -> bool {
-        self.options.open_noatime_enabled()
-    }
-
-    pub(super) const fn sparse_enabled(&self) -> bool {
-        self.options.sparse_enabled()
-    }
-
-    pub(super) const fn append_enabled(&self) -> bool {
-        self.options.append_enabled()
-    }
-
-    pub(super) const fn append_verify_enabled(&self) -> bool {
-        self.options.append_verify_enabled()
-    }
-
-    pub(super) const fn preallocate_enabled(&self) -> bool {
-        self.options.preallocate_enabled()
-    }
-
-    #[allow(dead_code)] // Accessor retained for future use; DeferredSync handles runtime selection
-    pub(super) const fn fsync_enabled(&self) -> bool {
-        self.options.fsync_enabled()
-    }
-
-    pub(super) const fn devices_enabled(&self) -> bool {
-        self.options.devices_enabled()
-    }
-
-    pub(super) const fn copy_devices_as_files_enabled(&self) -> bool {
-        self.options.copy_devices_as_files_enabled()
-    }
-
-    pub(super) const fn specials_enabled(&self) -> bool {
-        self.options.specials_enabled()
-    }
-
-    pub(super) const fn force_replacements_enabled(&self) -> bool {
-        self.options.force_replacements_enabled()
-    }
-
-    /// Returns the filter program used by xattr sync logic.
-    #[cfg(all(unix, feature = "xattr"))]
-    pub(super) const fn filter_program(
-        &self,
-    ) -> Option<&crate::local_copy::filter_program::FilterProgram> {
-        self.filter_program.as_ref()
-    }
-
-    /// Reports whether ACL preservation is enabled.
-    #[cfg(all(any(unix, windows), feature = "acl"))]
-    pub(super) const fn acls_enabled(&self) -> bool {
-        self.options.acls_enabled()
-    }
-
-    pub(super) const fn relative_paths_enabled(&self) -> bool {
-        self.options.relative_paths_enabled()
-    }
-
-    pub(super) const fn recursive_enabled(&self) -> bool {
-        self.options.recursive_enabled()
-    }
-
-    pub(super) const fn dirs_enabled(&self) -> bool {
-        self.options.dirs_enabled()
-    }
-
-    pub(super) const fn implied_dirs_enabled(&self) -> bool {
-        self.options.implied_dirs_enabled()
-    }
-
-    pub(super) const fn mkpath_enabled(&self) -> bool {
-        self.options.mkpath_enabled()
-    }
-
-    pub(super) const fn prune_empty_dirs_enabled(&self) -> bool {
-        self.options.prune_empty_dirs_enabled()
-    }
-
-    pub(super) const fn omit_dir_times_enabled(&self) -> bool {
-        self.options.omit_dir_times_enabled()
-    }
-
-    pub(super) const fn omit_link_times_enabled(&self) -> bool {
-        self.options.omit_link_times_enabled()
-    }
-
-    fn parent_relative_to_destination<'p>(&self, parent: &'p Path) -> Option<&'p Path> {
-        parent
-            .strip_prefix(self.destination_root())
-            .ok()
-            .filter(|path| !path.as_os_str().is_empty())
-    }
-
-    /// Dry-run replacement policy: remove destination (logically) and either
-    /// allow creation (Ok) or synthesize a "NotFound" error.
-    fn replace_parent_entry_dry_run(
-        &mut self,
-        parent: &Path,
-        existing: &fs::Metadata,
-        allow_creation: bool,
-    ) -> Result<(), LocalCopyError> {
-        if self.force_replacements_enabled() {
-            let relative = self.parent_relative_to_destination(parent);
-            self.force_remove_destination(parent, relative, existing)?;
-            if allow_creation {
-                Ok(())
-            } else {
-                Err(LocalCopyError::io(
-                    "create parent directory",
-                    parent.to_path_buf(),
-                    io::Error::from(io::ErrorKind::NotFound),
-                ))
-            }
-        } else {
-            Err(LocalCopyError::invalid_argument(
-                LocalCopyArgumentError::ReplaceNonDirectoryWithDirectory,
-            ))
-        }
-    }
-
-    /// Creation policy when creation is allowed and side effects are real
-    /// (non-dry-run): replace destination, create directory, and register progress.
-    fn replace_parent_entry_create(
-        &mut self,
-        parent: &Path,
-        existing: &fs::Metadata,
-    ) -> Result<(), LocalCopyError> {
-        if self.force_replacements_enabled() {
-            let relative = self.parent_relative_to_destination(parent);
-            self.force_remove_destination(parent, relative, existing)?;
-            fs::create_dir_all(parent).map_err(|error| {
-                LocalCopyError::io("create parent directory", parent, error)
-            })?;
-            self.register_progress();
-            Ok(())
-        } else {
-            Err(LocalCopyError::invalid_argument(
-                LocalCopyArgumentError::ReplaceNonDirectoryWithDirectory,
-            ))
-        }
-    }
-
-    /// Policy when creation is forbidden: replace destination and always return
-    /// the synthesized "NotFound" IO error (mirroring upstream behavior).
-    fn replace_parent_entry_forbidden(
-        &mut self,
-        parent: &Path,
-        existing: &fs::Metadata,
-    ) -> Result<(), LocalCopyError> {
-        if self.force_replacements_enabled() {
-            let relative = self.parent_relative_to_destination(parent);
-            self.force_remove_destination(parent, relative, existing)?;
-            Err(LocalCopyError::io(
-                "create parent directory",
-                parent.to_path_buf(),
-                io::Error::from(io::ErrorKind::NotFound),
-            ))
-        } else {
-            Err(LocalCopyError::invalid_argument(
-                LocalCopyArgumentError::ReplaceNonDirectoryWithDirectory,
-            ))
-        }
-    }
-
-    /// Ensures the parent directory exists, creating it if `--implied-dirs` or
-    /// `--mkpath` is enabled, or replacing a non-directory obstacle when
-    /// `--force` is set.
-    pub(super) fn prepare_parent_directory(&mut self, parent: &Path) -> Result<(), LocalCopyError> {
-        if parent.as_os_str().is_empty() {
-            return Ok(());
-        }
-
-        // Fast path: skip stat if this parent was already verified as a directory.
-        // With 10K files in one directory, this avoids 9,999 redundant statx calls.
-        if self.verified_parents.contains(parent) {
-            return Ok(());
-        }
-
-        let allow_creation = self.implied_dirs_enabled() || self.mkpath_enabled();
-        let keep_dirlinks = self.keep_dirlinks_enabled();
-
-        let result = if self.mode.is_dry_run() {
-            match fs::symlink_metadata(parent) {
-                Ok(existing) => {
-                    let ty = existing.file_type();
-                    if ty.is_dir() {
-                        Ok(())
-                    } else if keep_dirlinks && ty.is_symlink() {
-                        follow_symlink_metadata(parent).and_then(|metadata| {
-                            if metadata.file_type().is_dir() {
-                                Ok(())
-                            } else {
-                                self.replace_parent_entry_dry_run(parent, &existing, allow_creation)
-                            }
-                        })
-                    } else {
-                        self.replace_parent_entry_dry_run(parent, &existing, allow_creation)
-                    }
-                }
-                Err(error) if error.kind() == io::ErrorKind::NotFound => {
-                    if allow_creation {
-                        Ok(())
-                    } else {
-                        Err(LocalCopyError::io(
-                            "create parent directory",
-                            parent.to_path_buf(),
-                            error,
-                        ))
-                    }
-                }
-                Err(error) => Err(LocalCopyError::io(
-                    "inspect existing destination",
-                    parent.to_path_buf(),
-                    error,
-                )),
-            }
-        } else if allow_creation {
-            match fs::symlink_metadata(parent) {
-                Ok(existing) => {
-                    let ty = existing.file_type();
-                    if ty.is_dir() {
-                        Ok(())
-                    } else if keep_dirlinks && ty.is_symlink() {
-                        let metadata = follow_symlink_metadata(parent)?;
-                        if metadata.file_type().is_dir() {
-                            Ok(())
-                        } else {
-                            self.replace_parent_entry_create(parent, &existing)
-                        }
-                    } else {
-                        self.replace_parent_entry_create(parent, &existing)
-                    }
-                }
-                Err(error) if error.kind() == io::ErrorKind::NotFound => {
-                    fs::create_dir_all(parent).map_err(|error| {
-                        LocalCopyError::io("create parent directory", parent, error)
-                    })?;
-                    self.register_progress();
-                    Ok(())
-                }
-                Err(error) => Err(LocalCopyError::io(
-                    "create parent directory",
-                    parent.to_path_buf(),
-                    error,
-                )),
-            }
-        } else {
-            match fs::symlink_metadata(parent) {
-                Ok(existing) => {
-                    let ty = existing.file_type();
-                    if ty.is_dir() {
-                        Ok(())
-                    } else if keep_dirlinks && ty.is_symlink() {
-                        let metadata = follow_symlink_metadata(parent)?;
-                        if metadata.file_type().is_dir() {
-                            Ok(())
-                        } else {
-                            self.replace_parent_entry_forbidden(parent, &existing)
-                        }
-                    } else {
-                        self.replace_parent_entry_forbidden(parent, &existing)
-                    }
-                }
-                Err(error) => Err(LocalCopyError::io(
-                    "create parent directory",
-                    parent.to_path_buf(),
-                    error,
-                )),
-            }
-        };
-
-        if result.is_ok() {
-            self.verified_parents.insert(parent.to_path_buf());
-        }
-        result
-    }
-
-    pub(super) const fn remove_source_files_enabled(&self) -> bool {
-        self.options.remove_source_files_enabled()
-    }
-
-    pub(super) const fn compress_enabled(&self) -> bool {
-        self.options.compress_enabled()
-    }
-
-    /// Returns whether compression should be used for this file, considering
-    /// the skip-compress suffix list.
-    pub(super) fn should_compress(&self, relative: &Path) -> bool {
-        self.compress_enabled() && !self.options.should_skip_compress(relative)
-    }
-
-    pub(super) const fn compression_level(&self) -> CompressionLevel {
-        self.options.compression_level()
-    }
-
-    pub(super) const fn compression_algorithm(&self) -> CompressionAlgorithm {
-        self.options.compression_algorithm()
-    }
-
-    pub(super) const fn compression_threads(&self) -> Option<std::num::NonZeroU8> {
-        self.options.compression_threads()
-    }
-
-    pub(super) const fn block_size_override(&self) -> Option<NonZeroU32> {
-        self.options.block_size_override()
-    }
-
-    pub(super) const fn checksum_enabled(&self) -> bool {
-        self.options.checksum_enabled()
-    }
-
-    pub(super) const fn xxh64_dedup_enabled(&self) -> bool {
-        self.options.xxh64_dedup_enabled()
-    }
-
-    pub(super) const fn xxh64_dedup_size_limit(&self) -> u64 {
-        self.options.xxh64_dedup_size_limit()
-    }
-
-    pub(super) const fn size_only_enabled(&self) -> bool {
-        self.options.size_only_enabled()
-    }
-
-    pub(super) const fn ignore_times_enabled(&self) -> bool {
-        self.options.ignore_times_enabled()
-    }
-
-    pub(super) const fn ignore_existing_enabled(&self) -> bool {
-        self.options.ignore_existing_enabled()
-    }
-
-    pub(super) const fn existing_only_enabled(&self) -> bool {
-        self.options.existing_only_enabled()
-    }
-
-    pub(super) const fn ignore_missing_args_enabled(&self) -> bool {
-        self.options.ignore_missing_args_enabled()
-    }
-
-    pub(super) const fn delete_missing_args_enabled(&self) -> bool {
-        self.options.delete_missing_args_enabled()
-    }
-
-    pub(super) const fn update_enabled(&self) -> bool {
-        self.options.update_enabled()
-    }
-
-    pub(super) const fn partial_enabled(&self) -> bool {
-        self.options.partial_enabled()
-    }
-
-    pub(super) fn partial_directory_path(&self) -> Option<&Path> {
-        self.options.partial_directory_path()
-    }
-
-    pub(super) fn temp_directory_path(&self) -> Option<&Path> {
-        self.options.temp_directory_path()
-    }
-
-    pub(super) const fn inplace_enabled(&self) -> bool {
-        self.options.inplace_enabled()
-    }
-
-    #[cfg(unix)]
-    #[cfg(all(unix, feature = "xattr"))]
-    pub(super) const fn xattrs_enabled(&self) -> bool {
-        self.options.preserve_xattrs()
-    }
-
-    /// Evaluates filter rules to determine whether the entry is allowed for
-    /// transfer. Returns `true` if the entry passes all filters.
-    pub(super) fn allows(&self, relative: &Path, is_dir: bool) -> bool {
-        if let Some(program) = &self.filter_program {
-            let layers = self.dir_merge_layers.borrow();
-            let ephemeral = self.dir_merge_ephemeral.borrow();
-            let temp_layers = ephemeral.last().map(|entries| entries.as_slice());
-            program
-                .evaluate(
-                    relative,
-                    is_dir,
-                    layers.as_slice(),
-                    temp_layers,
-                    FilterContext::Transfer,
-                )
-                .allows_transfer()
-        } else if let Some(filters) = self.options.filter_set() {
-            filters.allows(relative, is_dir)
-        } else {
-            true
-        }
-    }
-
-    /// Returns `true` when a directory path is excluded by a non-directory-specific
-    /// filter rule.
-    ///
-    /// Used by the planner when `--prune-empty-dirs` is active: directories
-    /// excluded by generic patterns (e.g., `*`) should still be descended into
-    /// so that file-level include rules can be evaluated. Only directory-specific
-    /// exclude patterns (trailing `/`) should prevent traversal outright.
-    pub(super) fn excluded_dir_by_non_dir_rule(&self, relative: &Path) -> bool {
-        if let Some(program) = &self.filter_program {
-            let layers = self.dir_merge_layers.borrow();
-            let ephemeral = self.dir_merge_ephemeral.borrow();
-            let temp_layers = ephemeral.last().map(|entries| entries.as_slice());
-            program.excluded_dir_by_non_dir_rule(relative, layers.as_slice(), temp_layers)
-        } else if let Some(filters) = self.options.filter_set() {
-            filters.excluded_dir_by_non_dir_rule(relative)
-        } else {
-            false
-        }
-    }
-
-    /// Evaluates filter rules to determine whether a destination entry may be
-    /// deleted. Respects `--delete-excluded` when enabled.
-    pub(super) fn allows_deletion(&self, relative: &Path, is_dir: bool) -> bool {
-        let delete_excluded = self.options.delete_excluded_enabled();
-        if let Some(program) = &self.filter_program {
-            let layers = self.dir_merge_layers.borrow();
-            let ephemeral = self.dir_merge_ephemeral.borrow();
-            let temp_layers = ephemeral.last().map(|entries| entries.as_slice());
-            let outcome = program.evaluate(
-                relative,
-                is_dir,
-                layers.as_slice(),
-                temp_layers,
-                FilterContext::Deletion,
-            );
-            if delete_excluded {
-                outcome.allows_deletion() || outcome.allows_deletion_when_excluded_removed()
-            } else {
-                outcome.allows_deletion()
-            }
-        } else if let Some(filters) = self.options.filter_set() {
-            if delete_excluded {
-                filters.allows_deletion(relative, is_dir)
-                    || filters.allows_deletion_when_excluded_removed(relative, is_dir)
-            } else {
-                filters.allows_deletion(relative, is_dir)
-            }
-        } else {
-            true
-        }
-    }
-
     /// Access the batch writer for recording transfer operations.
     ///
     /// Returns a reference to the batch writer if batch mode is enabled,
@@ -1026,12 +509,6 @@ impl<'a> CopyContext<'a> {
     /// upstream's `flist_sort_and_clean()` produces after reading the batch.
     pub(super) fn record_batch_entry_sort_data(&mut self, name: &[u8], is_dir: bool) {
         self.batch_entry_sort_data.push((name.to_vec(), is_dir));
-    }
-
-    /// Returns whether `--numeric-ids` is enabled.
-    #[cfg(unix)]
-    pub(super) const fn numeric_ids_enabled(&self) -> bool {
-        self.options.numeric_ids_enabled()
     }
 }
